+++ src/main/java/org/sgrewritten/stargate/util/portal/PortalDestructionHelper.java
@@ -34,7 +34,7 @@ public final class PortalDestructionHelper {
      * @return <p>True if the destruction has been cancelled</p>
      */
     public static boolean destroyPortalIfHasPermissionAndCanPay(Player player, Portal portal, Runnable destroyAction, LanguageManager languageManager, StargateEconomyAPI economyManager) {
-        int cost = ConfigurationHelper.getInteger(ConfigurationOption.DESTROY_COST);
+        double cost = ConfigurationHelper.getDouble(ConfigurationOption.DESTROY_COST);
         StargatePermissionManager permissionManager = new StargatePermissionManager(player, languageManager);
 
         boolean hasPermission = permissionManager.hasDestroyPermissions((RealPortal) portal);
